+++ Providers/CrashlyticsProvider.m
@@ -25,7 +25,7 @@
     }
 
     if (email) {
-        [Crashlytics setUserName:email];
+        [Crashlytics setUserEmail:email];
     }
 }
 
@@ -34,7 +34,14 @@
 }
 
 - (void)event:(NSString *)event withProperties:(NSDictionary *)properties {
-    CLSLog(@"%@", event);
+    NSString *log;
+    if (properties) {
+        log = [NSString stringWithFormat:@"%@%@", event, properties];
+    } else {
+        log = event;
+    }
+    
+    CLSLog(@"%@", log);
 }
 
 - (void)remoteLog:(NSString *)parsedString {